Passport -> firebase

+++ src/client/main.tsx
@@ -10,7 +10,7 @@ import * as injectTapEventPlugin from 'react-tap-event-plugin';
 injectTapEventPlugin();
 
 import {EventEmitter} from 'eventemitter3';
-import * as Passport from 'passport';
+// import * as Passport from 'passport';
 
 import IconButton from 'material-ui/IconButton';
 import IconMenu from 'material-ui/IconMenu';
@@ -38,7 +38,7 @@ import Timeline from './timeline';
 import SearchBox from './components/SearchBox';
 import DialogBox from './components/DialogBox';
 
-import {AppState} from './app';
+import {AppState, UserProfile} from './app';
 import * as common from './common';
 import Constant from './constant';
 // import {UserInfo} from './app';
@@ -64,8 +64,8 @@ export interface MainState {
   contents: number,
   wordList: WordInfo[],
   searchWord: string,
-//  userInfo: UserInfo,
-  profile: Passport.Profile,
+  profile: UserProfile,
+  database: firebase.database.Database,
   deleteDialogFlag: boolean,
   addmylistDialogFlag: boolean,
   onDeleteItem: ()=>void,
@@ -90,12 +90,12 @@ export default class Main extends React.Component<AppState, MainState>{
         userId: null,
         userName: null,
         icon: null,
-        createDate: +new Date(),
-        updateDate: +new Date(),
+        createDate: +new Date('1989-01-01 00:00:00'),
+        updateDate: +new Date('1989-01-01 00:00:00'),
       }],
       searchWord: '',
-      // userInfo: this.props.profile,
       profile: this.props.profile,
+      database: null,
       onDeleteItem: ()=>{},
       onAddMylist: ()=>{},
       deleteDialogFlag: false,
@@ -103,26 +103,55 @@ export default class Main extends React.Component<AppState, MainState>{
     };
 
     this.state.emitter.on('cookieRegister', (kv: KeyValueItem)=>{
-      const kvtext: string = kv.key + Constant.SEPARATOR + kv.value + Constant.SEPARATOR + Constant.HASHTAG;
-      API.postTweet(kvtext)
-      .then((response: any)=>{
-        // console.log('response',response);
-        if(response && response.statusCode === 200) {
-          console.log('success!');
-          location.reload();
-        }
-      }).catch((err: string)=>{
-        console.log(err);
-      });
+      // console.log('register profile', this.props.profile);
+
+      if(this.props.profile.provider === 'twitter.com') {
+        const kvtext: string = kv.key + Constant.SEPARATOR + kv.value + Constant.SEPARATOR + Constant.HASHTAG;
+        API.postTweet(kvtext)
+        .then((response: any)=>{
+          // console.log('response',response);
+          if(response && response.statusCode === 200) {
+            console.log('success!');
+            location.reload();
+          }
+        }).catch((err: string)=>{
+          console.log(err);
+        });
+      }
+      else if(this.props.profile.provider === 'firebase') {
+        const testUserPath: string = 'users/'+this.props.profile.id+'/';
+        const keyid: string = this.state.database.ref().child(testUserPath).push().key;
+        const newData: WordInfo = {
+          id: keyid,
+          key: kv.key,
+          value: kv.value,
+          userId: this.state.profile.id,
+          userName: this.state.profile.displayName,
+          icon: this.state.profile.photoURL,
+          createDate: +new Date(),
+          updateDate: +new Date(),        
+        };
+        this.state.database.ref(testUserPath + keyid).set(newData)
+        .then((res) => {
+          console.log(res);
+        }).catch((err) => {
+          console.log(err);
+        });
+      }
+
     }).on('cookieSearch', (searchWord: string)=> {
       this.setState({searchWord});
+
     }).on('cookieItemDelete', (id: string)=> {
       //delete
       const wordlist: WordInfo[] = this.state.wordList;
-      let deleteIdx: number = null;
+      let deleteIdx: number = null, target: WordInfo = null;
       wordlist.forEach((w: WordInfo, i: number)=>{
         if(w.id === id){
           deleteIdx = i;
+          target = w;
+          target.value = null;
+          target.updateDate = +new Date();
         }
       });
       if(typeof deleteIdx === 'number'){
@@ -130,12 +159,29 @@ export default class Main extends React.Component<AppState, MainState>{
           deleteDialogFlag: true,
           onDeleteItem: ()=>{
             wordlist.splice(deleteIdx, 1);
-            API.deleteItem(id).then(()=>{
-              this.setState({wordList: wordlist, deleteDialogFlag: false});
-            }).catch((err)=>{
-              alert(err);
-              this.setState({deleteDialogFlag: false});
-            });
+
+            if(this.props.profile.provider === 'twitter.com') {
+              API.deleteItem(id).then(()=>{
+                this.setState({wordList: wordlist, deleteDialogFlag: false});
+              }).catch((err)=>{
+                alert(err);
+                this.setState({deleteDialogFlag: false});
+              });
+            }
+            else if(this.props.profile.provider === 'firebase'){
+              this.props.fb.database().ref('/users/'+this.props.profile.id)
+              .set(target)
+              .then((result) => {
+                console.log(result);
+                this.setState({wordList: wordlist, deleteDialogFlag: false});
+              })
+              .catch((err) => {
+                console.log(err);
+                alert(err);
+                this.setState({deleteDialogFlag: false});
+              });
+            }
+
           }
         });
       }
@@ -144,42 +190,97 @@ export default class Main extends React.Component<AppState, MainState>{
   }
 
   componentDidMount() {
-    //ユーザ情報取得
-    this.setState({contents: 0});
-    API.getCredentials()
-    .then((response: {result: boolean, err: any})=>{
-      // console.log('getCredentials',result);
-      if(!response.result){
-        throw response.err;
-      }
-      return API.getUserTimeline();
-    }).then((response: API.TweetInfo[])=>{
-      // console.log(response);
-      if(response){
-        const wordList: WordInfo[] = response
-        .filter((v)=>{
-          return v.text.indexOf(Constant.HASHTAG) > 0 && v.text.indexOf(Constant.ATTMARK) !== 0;
-        })
-        .map((v)=>{
-          const keyValues = v.text.split(Constant.SEPARATOR);
-          return {
-            id: v.id_str,
-            key: keyValues[0],
-            value: keyValues[1],
-            userId: v.user.name,
-            userName: v.user.screen_name,
-            icon: v.user.profile_image_url,
-            createDate: +new Date(v.created_at),
-            updateDate: +new Date(v.created_at),
-          };
+    // firebase Database
+    const fbDB: firebase.database.Database = this.props.fb.database();
+    this.setState({
+      contents: 0,
+      database: fbDB,
+    });
+
+    let wordList: WordInfo[] = [];
+
+    //mccookite db data
+    const mcctestRef: firebase.database.Reference = fbDB.ref('users/mcctest');
+    mcctestRef.once('value', (snapshot) => {
+      const values: any = snapshot.val();
+      if(values) {
+        const vKeyids: string[] = Object.keys(values);
+        const preWordList: WordInfo[] = this.state.wordList;
+        const nextWordList: WordInfo[] = vKeyids.map((keyid) => {
+          return values[keyid];
         });
-        this.setState({
-          wordList,
-          contents: 0
+
+        wordList = preWordList.concat(nextWordList)
+        .sort((w1: WordInfo, w2: WordInfo)=>w2.updateDate - w1.updateDate);
+
+        this.setState({wordList: wordList});
+      }
+
+      //ユーザ情報取得
+      if(this.state.profile.provider === 'twitter.com') {
+        API.getCredentials()
+        .then((response: {result: boolean, err: any})=>{
+          // console.log('getCredentials',result);
+          if(!response.result){
+            throw response.err;
+          }
+          return API.getUserTimeline();
+        }).then((response: API.TweetInfo[])=>{
+          if(response){
+            const preWordList: WordInfo[] = this.state.wordList;
+            const nextWordList: WordInfo[] = response
+            .filter((v)=>{
+              return v.text.indexOf(Constant.HASHTAG) > 0 && v.text.indexOf(Constant.ATTMARK) !== 0;
+            })
+            .map((v)=>{
+              const keyValues = v.text.split(Constant.SEPARATOR);
+              return {
+                id: v.id_str,
+                key: keyValues[0],
+                value: keyValues[1],
+                userId: v.user.name,
+                userName: v.user.screen_name,
+                icon: v.user.profile_image_url,
+                createDate: +new Date(v.created_at),
+                updateDate: +new Date(v.created_at),
+              };
+            });
+            wordList = preWordList.concat(nextWordList)
+            .sort((w1: WordInfo, w2: WordInfo)=>w2.updateDate - w1.updateDate);
+            this.setState({wordList: wordList});
+          }
+        }).catch((err)=>{
+          console.log(err);
+        });
+
+      } else if(this.state.profile.provider === 'firebase') {
+        //TODO my wordlist
+        const mcctestRef: firebase.database.Reference = fbDB.ref('users/' + this.state.profile.id);
+        mcctestRef.on('value', (snapshot) => {
+          const values: any = snapshot.val();
+          if(values) {
+            const vKeyids: string[] = Object.keys(values);
+            const preWordList: WordInfo[] = this.state.wordList;
+            const nextWordList: WordInfo[] = vKeyids
+            .filter((kid) => {
+              let _notExist: boolean = true;
+              preWordList.forEach((w: WordInfo) => {
+                if(w.id === kid) {
+                  _notExist = false;
+                }
+              });
+              return _notExist;
+            })
+            .map((keyid) => {
+              return values[keyid];
+            });
+            wordList = preWordList.concat(nextWordList)
+            .sort((w1: WordInfo, w2: WordInfo)=>w2.updateDate - w1.updateDate);
+            this.setState({wordList: wordList});
+          }
         });
       }
-    }).catch((err)=>{
-      console.log(err);
+
     });
   }
 
@@ -188,7 +289,7 @@ export default class Main extends React.Component<AppState, MainState>{
     this.setState({contents: 0});
   }
   handleLogout() {
-    // this.props.onLogout(this.state.userId);
+    this.props.onLogout(this.state.profile.id);
   }
 
   render() {
